change totalReward to BigInt

+++ backend/src/main/kotlin/com/retypeme/project/statistic/model/StatisticModel.kt
@@ -1,12 +1,13 @@
 package com.retypeme.project.statistic.model
 
 import java.io.Serializable
+import java.math.BigInteger
 
 data class StatisticModel(
     val userId: String,
     val completedDuels: Int,
     val averageSpeed: Double,
-    val totalReward: Int,
+    val totalReward: BigInteger,
     val overallWinsInDuels: Int,
     val maxSpeed: Double,
     val topSpeeds: List<Double>
